requirements.txt
feedparser
pandas
beautifulsoup4
nltk
vaderSentiment
transformers
torch
scipy
streamlit
plotly
yfinance
requests
python-dotenv
apscheduler
python-dateutil
tqdm
scikit-learn
joblib
PyYAML